Оптимизировал очистку используемых объектов при выходе из программы. Оптимизировал доступ к блокам карты за счет создания карты при входе в УО или смене карты. Исправил баг с не корректным получением данных с карт индексом выше 1.

+++ MultiMap.cpp
@@ -91,7 +91,7 @@ void TMultiMap::LoadMap(TGumpMap *gump)
 
 		BYTE size = pic & 0x7F;
 
-		DWORD color = 0;
+		WORD color = 0;
 		if (pic & 0x80)
 			color = data_color;
